fix build for the two SipHash variations
fixes #16471
